"Color Game 2019!"
You need to pick the right RGB color which is showing on the top! if you choose the correct one, the color of the top will be changed as a correct color!

There are two modes(Hard/Easy), so you can click one of the modes, and enjoy the game!
You will learn what RGB color means while playing the game!